requirements.txt
PyMySQL==1.0.2
cryptography==3.4.7
Flask==1.1.2
flask_sqlalchemy==2.5.1
sqlalchemy < 1.4.0
flask_migrate==2.7.0
flask_script==2.0.6
sqlalchemy_utils==0.36.8
alembic~=1.6.5
pytz~=2021.1
python-dateutil~=2.8.1
numpy~=1.21.0
redis~=3.5.3
flask_jwt_extended==4.1.0